add: Script load requirements
This commit adds "environment", "depends" and "breaks" requirements to
the script manifest file. These requirements work similar to
fabric.mod.json requirements. All requirements must be met for a script
to be loaded.

This commit got rid of the ability to define script manifests in mods
fabric.mod.json files, because it would require a completely separate
deserializer. An alternative to this may be to look for "allium" folders
inside mod jars which could include full scripts (manifest + lua files)
instead!